backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
numpy==1.21.6
opencv-python==4.5.5.64
pillow==9.5.0
torch==1.12.1
torchvision==0.13.1
scikit-learn==1.0.2
matplotlib==3.5.3
sqlalchemy==1.4.46
python-dotenv==0.21.0
boto3==1.26.137
python-multipart==0.0.6
pydantic==1.10.4
